latinization moved to main script
